+++ core-components/terraform/layers/opscenter-resources/_outputs.tf
@@ -1,3 +1,0 @@
-output "opscenter_elb" {
-  value = module.opscenter.opscenter_elb
-}