+++ core-components/terraform/modules/cassandra/templates.tf
@@ -1,16 +1,13 @@
-data "template_file" "dse-init" {
-  template = file("${path.module}/files/dse-init.tpl")
+data "template_file" "cassandra-init" {
+  template = file("${path.module}/files/cassandra-init.tpl")
 
   vars = {
-    dc_name        = var.datacenter
-    auto_start_dse = var.auto_start_dse
-    region         = var.region
-    ssh_bucket     = var.tfstate_bucket
-    ssh_prefix     = "${local.cluster_key}/files/ssh/ec2-user/user-keys.yaml"
-    graph_enabled  = var.graph_enabled
-    solr_enabled   = var.solr_enabled
-    spark_enabled  = var.spark_enabled
-    ec2_tag_map    = jsonencode(merge(var.ec2_tags, local.required_ec2_tags))
+    dc_name              = var.datacenter
+    auto_start_cassandra = var.auto_start_cassandra
+    region               = var.region
+    ssh_bucket           = var.tfstate_bucket
+    ssh_prefix           = "${local.cluster_key}/files/ssh/ec2-user/user-keys.yaml"
+    ec2_tag_map          = jsonencode(merge(var.ec2_tags, local.required_ec2_tags))
   }
 }
 
@@ -26,9 +23,9 @@ data "template_cloudinit_config" "cassandra" {
   }
 
   part {
-    filename     = "dse-init.sh"
+    filename     = "cassandra-init.sh"
     content_type = "text/x-shellscript"
-    content      = data.template_file.dse-init.rendered
+    content      = data.template_file.cassandra-init.rendered
   }
 }