+++ core-components/terraform/modules/iam-resources/bastion-role.tf
@@ -24,8 +24,8 @@ resource "aws_iam_policy" "bastion-ssm-policy" {
 }
 
 resource "aws_iam_role_policy_attachment" "bastion-ssm-attach" {
-  role        = aws_iam_role.bastion-role.name
-  policy_arn  = aws_iam_policy.bastion-ssm-policy.arn
+  role       = aws_iam_role.bastion-role.name
+  policy_arn = aws_iam_policy.bastion-ssm-policy.arn
 }
 
 ###################
@@ -59,12 +59,12 @@ data "aws_iam_policy_document" "bastion-bootstrap-doc" {
 }
 
 resource "aws_iam_policy" "bastion-bootstrap-policy" {
-  depends_on  = [aws_iam_role.bastion-role]
-  name        = "${var.prefix}bastion-bootstrap-policy${var.suffix}"
-  policy      = data.aws_iam_policy_document.bastion-bootstrap-doc.json
+  depends_on = [aws_iam_role.bastion-role]
+  name       = "${var.prefix}bastion-bootstrap-policy${var.suffix}"
+  policy     = data.aws_iam_policy_document.bastion-bootstrap-doc.json
 }
 
 resource "aws_iam_role_policy_attachment" "bastion-bootstrap-attach" {
-  role        = aws_iam_role.bastion-role.name
-  policy_arn  = aws_iam_policy.bastion-bootstrap-policy.arn
+  role       = aws_iam_role.bastion-role.name
+  policy_arn = aws_iam_policy.bastion-bootstrap-policy.arn
 }