+++ core-components/terraform/modules/opscenter/amis.tf
@@ -1,9 +1,0 @@
-data "aws_ami" "opscenter" {
-  most_recent = true
-  owners      = ["${var.ami_owner_id}"]
-
-  filter {
-    name   = "name"
-    values = ["${var.ami_prefix}*"]
-  }
-}