+++ core-components/terraform/modules/vpc-create/_outputs.tf
@@ -1,23 +1,23 @@
 output "vpc_id" {
-  value = aws_vpc.dse-vpc.id
+  value = aws_vpc.cassandra-vpc.id
 }
 
 output "vpc_cidr" {
-  value = aws_vpc.dse-vpc.cidr_block
+  value = aws_vpc.cassandra-vpc.cidr_block
 }
 
 output "data_subnet_ids" {
-  value = aws_subnet.dse-vpc-data.*.id
+  value = aws_subnet.cassandra-vpc-data.*.id
 }
 
 output "data_subnet_cidr_blocks" {
-  value = aws_subnet.dse-vpc-data.*.cidr_block
+  value = aws_subnet.cassandra-vpc-data.*.cidr_block
 }
 
 output "ingress_subnet_ids" {
-  value = aws_subnet.dse-vpc-ingress.*.id
+  value = aws_subnet.cassandra-vpc-ingress.*.id
 }
 
 output "ingress_subnet_cidr_blocks" {
-  value = aws_subnet.dse-vpc-ingress.*.cidr_block
+  value = aws_subnet.cassandra-vpc-ingress.*.cidr_block
 }